Migrate onboarding_presentation to KMP (Phase 9e) - completes Phase 9
Last of the 8 *_presentation modules, sequenced last per the migration
plan since it's the one with genuine platform-specific coupling: the
Welcome screen's sign-in button needs an Activity to present Google's
credential UI, and AuthService.signIn(activity: Activity) still leaks
that Android type - deferred to the iOS phase per the plan (presenting
sign-in UI is one of the few truly platform-specific things in this app).

Split core/domain/auth/service/AuthService.kt (previously one androidMain
file) rather than leaving the whole onboarding_presentation module on
Android:
- AuthService (the Activity-taking interface) stays androidMain-only.
- AuthServiceResult/AuthServiceUserData - pure data types with zero
  Android imports, they only happened to live in the same file - moved to
  a new core/commonMain/.../AuthServiceResult.kt. Same package name in
  both files, so no import changes needed anywhere that already
  referenced them (GoogleAuthService.kt in :app, etc.) - Gradle/Kotlin
  resolves the split transparently.

This unblocked AuthEvent.kt, SignInState.kt, SignInViewModel.kt, and
di/Module.kt (all of which depended on AuthServiceResult, not AuthService
itself) to move to commonMain. Only OnBoardingScreen.kt - the one file
that actually calls authService.signIn(context as Activity) - stays in
onboarding_presentation's androidMain, alongside its own AndroidManifest
and test source sets. Same koinViewModel Android->multiplatform swap
applied there too.

base_module_compose -> base_kmp_compose_module.

This completes Phase 9: all 8 *_presentation modules (grades_setting,
quick, subject_settings, gpa_system_settings, semester_marks,
semester_subjctets, semester_history, onboarding) are now on Kotlin
Multiplatform.

Note on verification: unverified in this sandbox - no Android SDK, Gradle
9.3.1 download blocked by network policy (only 8.14.3 available). Needs a
real build plus a manual sign-in/sign-out smoke test end to end (this is
the screen most tightly coupled to real Firebase Auth + Credential
Manager behavior).
